Fix the environment variables
Inside podman, the variables set inside the .env.* files are set as-is,
as pure strings, in the environment memory space of the process and
result in unusable environment variables.

They are visible throught the 'env' tool but not usable by any shells
as they contains spaces between the equal sign.

Moreover, for the same resaon, the quotes around the values are
useless.  Better to not set them there and set them at the time we use
it if required.